Move ParseException to top level exception package
