+++ src/main/java/seedu/address/logic/parser/Parser.java
@@ -1,7 +1,7 @@
 package seedu.address.logic.parser;
 
+import seedu.address.exception.ParseException;
 import seedu.address.logic.commands.Command;
-import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
  * Represents a Parser that is able to parse user input into a {@code Command} of type {@code T}.
