+++ src/main/java/seedu/address/ui/CommandBox.java
@@ -5,8 +5,8 @@ import javafx.fxml.FXML;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.Region;
 import seedu.address.exception.CommandException;
+import seedu.address.exception.ParseException;
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
  * The UI component that is responsible for receiving user command inputs.
