+++ src/test/java/seedu/address/logic/parser/CommandParserTestUtil.java
@@ -2,8 +2,8 @@ package seedu.address.logic.parser;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import seedu.address.exception.ParseException;
 import seedu.address.logic.commands.Command;
-import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
  * Contains helper methods for testing command parsers.
